+++ crates/editor/src/workspace/mod.rs
@@ -265,6 +265,7 @@ impl Workspace {
         }
     }
 
+    #[profiling::function]
     pub fn show(
         &mut self,
         ui: &mut egui::Ui,